+++ src/Pen/pen.js
@@ -1,5 +1,5 @@
 import { moduleName } from "../../MaterialPlane.js";
-import { cursor, findToken, scaleIRinput } from "../Misc/misc.js";
+import { cursor, findToken, scaleIRinput, compatibleCore } from "../Misc/misc.js";
 import { IRtoken } from "../IRtoken/IRtoken.js";
 import { penMenu } from "./penMenu.js";
 
@@ -97,11 +97,12 @@ export class Pen extends CanvasLayer {
                     color: "0x00FF00",
                     rawCoords: coords
                 });
+
                 if (this.oldCommand != 'penIdle')   {
                     status = 'release';
                     this.holdTime = undefined;
                 }
-                else if (this.drawing != undefined && this.drawing.data.type == CONST.DRAWING_TYPES.POLYGON && this.newDrawing) {
+                else if (this.drawing != undefined && (compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) == CONST.DRAWING_TYPES.POLYGON && this.newDrawing) {
                     const event = {
                         data: {
                             destination: {
@@ -344,29 +345,36 @@ export class Pen extends CanvasLayer {
                 else if (this.menu.selectedDrawingTool == 6) {
                     this.drawing = this.getNearestDrawing(data);
                     if (this.drawing != undefined) {
-                        this.drawing.delete();
+                        this.drawing.document.delete();
                         this.drawing = undefined;
                         this.lastDrawing = undefined;
                     }
                 }
-                else if (this.drawing != undefined && this.drawing.data.type == CONST.DRAWING_TYPES.POLYGON && this.newDrawing) {
+                else if (this.drawing != undefined && (compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) == CONST.DRAWING_TYPES.POLYGON && (compatibleCore('10.0') ? this.drawing.document.bezierFactor == 0 : true) && this.newDrawing) {
                     this.drawing._addPoint({x:data.x,y:data.y},false)
                 }
                 else {
                     let tool = this.menu.selectedDrawingToolName;
                     let drawingData = canvas.layers.filter(layer => layer.name == 'DrawingsLayer')[0]._getNewDrawingData({x:data.x,y:data.y})
-                    
                     let type;
                     if (tool == 'rect') type = CONST.DRAWING_TYPES.RECTANGLE;
                     else if (tool == 'ellipse') type = CONST.DRAWING_TYPES.ELLIPSE;
                     else if (tool == 'polygon') type = CONST.DRAWING_TYPES.POLYGON;
-                    else if (tool == 'freehand') type = CONST.DRAWING_TYPES.FREEHAND;
+                    else if (tool == 'freehand') {
+                        type = compatibleCore('10.0') ? CONST.DRAWING_TYPES.POLYGON : CONST.DRAWING_TYPES.FREEHAND;
+                        ui.controls.controls.find(c => c.name == 'drawings').activeTool = 'freehand';
+                        ui.controls.render(); 
+                    }
 
-                    drawingData.type = type;
+                    if (compatibleCore('10.0')) {
+                        drawingData.shape.type = type;
+                        if (tool == 'freehand') drawingData.bezierFactor = 0.5;
+                    }
+                    else drawingData.type = type;
                     drawingData.fillColor = this.menu.colors[this.menu.selectedFillColor].hex;
-                    drawingData.fillType = this.menu.colors[this.menu.selectedFillColor].hex == 'none' ? 0 : 1;
+                    drawingData.fillType = this.menu.colors[this.menu.selectedFillColor].name == 'none' ? 0 : 1;
                     drawingData.strokeColor = this.menu.colors[this.menu.selectedLineColor].hex;
-                    drawingData.strokeAlpha = this.menu.colors[this.menu.selectedLineColor].hex == 'none' ? 0 : 1;
+                    drawingData.strokeAlpha = this.menu.colors[this.menu.selectedLineColor].name == 'none' ? 0 : 1;
                     drawingData.x = data.x;
                     drawingData.y = data.y;
                     drawingData.strokeAlpha = 1;
@@ -375,18 +383,16 @@ export class Pen extends CanvasLayer {
                     const document = new DrawingDocument(drawingData, {parent: canvas.scene});
                     this.drawing = new Drawing(document);
                     canvas.drawings.preview.addChild(this.drawing);
-                    this.drawing._addPoint({x:data.x,y:data.y},false)
+                    this.drawing._addPoint({x:data.x, y:data.y},false)
                     this.drawing.draw();
                     
-    
                     this.newDrawing = true;
                 }
             }
             else if (status == 'hold') {
                 if (this.drawing == undefined || this.menu.selectedDrawing == 6) return;
-                
-                if ((this.drawing.data.type == CONST.DRAWING_TYPES.POLYGON || this.drawing.data.type == CONST.DRAWING_TYPES.FREEHAND) && this.newDrawing) {
 
+                if (((compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) == CONST.DRAWING_TYPES.POLYGON || (compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) == CONST.DRAWING_TYPES.FREEHAND) && this.newDrawing) {
                     const event = {
                         data: {
                             destination: {
@@ -404,9 +410,15 @@ export class Pen extends CanvasLayer {
                 else {
                     const dx = data.x - this.drawing.x;
                     const dy = data.y - this.drawing.y;
-    
-                    this.drawing.data.width = dx;
-                    this.drawing.data.height = dy;
+                    
+                    if (compatibleCore('10.0')) {
+                        this.drawing.document.shape.width = dx;
+                        this.drawing.document.shape.height = dy;
+                    }
+                    else {
+                        this.drawing.data.width = dx;
+                        this.drawing.data.height = dy;
+                    }
                     this.drawing.refresh();
                 }
                 
@@ -414,17 +426,23 @@ export class Pen extends CanvasLayer {
             else if (status == 'release') {
                 if (this.drawing == undefined || this.menu.selectedDrawing == 6) return;
                 
-                if (this.drawing.data.type != CONST.DRAWING_TYPES.POLYGON) {
+                if ((compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) != CONST.DRAWING_TYPES.POLYGON || (compatibleCore('10.0') ? this.drawing.document.bezierFactor > 0 : false)) {
                     if (this.newDrawing) {
                         this.newDrawing = false;
                         canvas.drawings.preview.removeChild(this.drawing);
                         const cls = getDocumentClass('Drawing');
-                        this.lastDrawing = await cls.create(this.drawing.data.toObject(false), {parent: canvas.scene});
+                        if (compatibleCore('10.0')) {
+                            this.lastDrawing = await cls.create(this.drawing.document.toObject(false), {parent: canvas.scene});
+                            this.drawing.destroy();
+                        }
+                        else this.lastDrawing = await cls.create(this.drawing.data.toObject(false), {parent: canvas.scene});
+                        
                         this.drawing == undefined;
                     }
                     else {   
-                        this.drawing.update({width:this.drawing.data.width,height:this.drawing.data.height});
+                        this.drawing.document.update({width:this.drawing.shape.width, height:this.drawing.shape.height});
                         this.lastDrawing = this.drawing;
+                        this.drawing.refresh();
                         this.drawing == undefined;
                     }
                 }
@@ -432,11 +450,12 @@ export class Pen extends CanvasLayer {
         }
         else if (command == 'penLeft') {
             if (status == 'click') {
-                if (this.newDrawing && this.drawing.data.type == CONST.DRAWING_TYPES.POLYGON) {
+                if (this.newDrawing && (compatibleCore('10.0') ? this.drawing.type : this.drawing.data.type) == CONST.DRAWING_TYPES.POLYGON) {
                     this.newDrawing = false;
                     canvas.drawings.preview.removeChild(this.drawing);
                     const cls = getDocumentClass('Drawing');
-                    this.lastDrawing = await cls.create(this.drawing.data.toObject(false), {parent: canvas.scene});
+                    if (compatibleCore('10.0')) this.lastDrawing = await cls.create(this.drawing.document.toObject(false), {parent: canvas.scene});
+                    else this.lastDrawing = await cls.create(this.drawing.data.toObject(false), {parent: canvas.scene});
                     this.drawing == undefined;
                 }
                 else {
@@ -446,13 +465,19 @@ export class Pen extends CanvasLayer {
             }
             else if (status == 'hold') {
                 if (this.drawing == undefined || this.menu.selectedDrawing == 6) return;
-                this.drawing.data.x = data.x - this.drawing.data.width/2;
-                this.drawing.data.y = data.y - this.drawing.data.height/2;
+                if (compatibleCore('10.0')) {
+                    this.drawing.document.x = data.x - this.drawing.shape.width/2;
+                    this.drawing.document.y = data.y - this.drawing.shape.height/2;
+                }
+                else {
+                    this.drawing.data.x = data.x - this.drawing.data.width/2;
+                    this.drawing.data.y = data.y - this.drawing.data.height/2;
+                }
                 this.drawing.refresh();
             }
             else if (status == 'release') {
                 if (this.drawing == undefined || this.menu.selectedDrawing == 6) return;
-                this.drawing.update({x:this.drawing.data.x,y:this.drawing.data.y})
+                this.drawing.document.update({x:this.drawing.x, y:this.drawing.y});
             }
         }
         else if (command == 'penRear') {
@@ -500,7 +525,7 @@ export class Pen extends CanvasLayer {
                 else if (this.menu.selectedTemplate == 6) {
                     this.template = this.getNearestTemplate(data);
                     if (this.template != undefined) {
-                        this.template.delete();
+                        this.template.document.delete();
                         this.template = undefined;
                     }
                 }
@@ -517,7 +542,7 @@ export class Pen extends CanvasLayer {
                         y: snappedPosition.y,
                         distance: 1,
                         direction: 0,
-                        fillColor: game.user.data.color || "#FF0000"
+                        fillColor: (compatibleCore('10.0') ? game.user.color : game.user.data.color) || "#FF0000"
                     };
     
                     // Apply some type-specific defaults
@@ -540,9 +565,15 @@ export class Pen extends CanvasLayer {
                 const dx = data.x - this.template.x;
                 const dy = data.y - this.template.y;
                 const length = Math.round(Math.sqrt(dx*dx + dy*dy)*canvas.dimensions.distance/canvas.dimensions.size);
-                this.template.data.distance = length;
                 const angle = 90-Math.atan2(dx,dy)*180/Math.PI;
-                this.template.data.direction = angle;
+                if (compatibleCore('10.0')) {
+                    this.template.document.distance = length;
+                    this.template.document.direction = angle;
+                }
+                else {
+                    this.template.data.distance = length;
+                    this.template.data.direction = angle;
+                }
                 this.template.refresh();
                 this.template.highlightGrid();
             }
@@ -551,12 +582,16 @@ export class Pen extends CanvasLayer {
                 if (this.newTemplate) {
                     this.menu.setSelected(1,true)
                     this.newTemplate = false;
+                    this.template.controlIcon.visible = false;
                     canvas.templates.preview.removeChild(this.template);
                     const cls = getDocumentClass('MeasuredTemplate');
-                    return cls.create(this.template.data.toObject(false), {parent: canvas.scene});
+                    if (compatibleCore('10.0')) return cls.create(this.template.document.toObject(false), {parent: canvas.scene});
+                    else return cls.create(this.template.data.toObject(false), {parent: canvas.scene});
                 }
-                else    
-                    this.template.update({distance:this.template.data.distance,direction:this.template.data.direction})
+                else {
+                    this.template.document.update({distance:this.template.document.distance,direction:this.template.document.direction});
+                }  
+                    
             }   
         }
         else if (command == 'penLeft') {
@@ -564,15 +599,21 @@ export class Pen extends CanvasLayer {
                 this.template = this.getNearestTemplate(data);
             else if (status == 'hold') {
                 if (this.template == undefined || this.menu.selectedTemplate == 6) return;
-                this.template.data.x = data.x;
-                this.template.data.y = data.y;
+                if (compatibleCore('10.0')) {
+                    this.template.document.x = data.x;
+                    this.template.document.y = data.y;
+                }
+                else {
+                    this.template.data.x = data.x;
+                    this.template.data.y = data.y;
+                }
                 this.template.refresh();
                 this.template.highlightGrid();
             }
             else if (status == 'release') {
                 if (this.template == undefined || this.menu.selectedTemplate == 6) return;
                 const snappedPosition = canvas.grid.getSnappedPosition(data.x, data.y, canvas.templates.gridPrecision);
-                this.template.update({x:snappedPosition.x,y:snappedPosition.y})
+                this.template.document.update({x:snappedPosition.x,y:snappedPosition.y});
             }
         }
         else if (command == 'penRear') {
@@ -598,9 +639,17 @@ export class Pen extends CanvasLayer {
                     length:data.length,
                     angle:data.angle
                 }
-                if (!isNaN(angleChange)) this.template.data.direction += angleChange;
-                this.template.data.x = data.x;
-                this.template.data.y = data.y;
+                
+                if (compatibleCore('10.0')) {
+                    if (!isNaN(angleChange)) this.template.document.direction += angleChange;
+                    this.template.document.x = data.x;
+                    this.template.document.y = data.y;
+                }
+                else {
+                    if (!isNaN(angleChange)) this.template.data.direction += angleChange;
+                    this.template.data.x = data.x;
+                    this.template.data.y = data.y;
+                }
                 
                 this.template.refresh();
                 this.template.highlightGrid();
@@ -608,7 +657,7 @@ export class Pen extends CanvasLayer {
             else if (status == 'release') {
                 if (this.template == undefined || this.menu.selectedTemplate == 6) return;
                 const snappedPosition = canvas.grid.getSnappedPosition(data.x, data.y, canvas.templates.gridPrecision);
-                this.template.update({x:snappedPosition.x,y:snappedPosition.y,direction:this.template.data.direction})
+                this.template.document.update({x:snappedPosition.x,y:snappedPosition.y,direction:this.template.document.direction});
             }  
         }
     }
@@ -686,7 +735,9 @@ export class Pen extends CanvasLayer {
         let nearestDrawing = undefined;
         let nearestDistance = 10000;
         for (let drawing of drawings) {
-            const dist = Math.abs(drawing.x+drawing.data.width/2-data.x) + Math.abs(drawing.y+drawing.data.height/2-data.y)
+            let dist;
+            if (compatibleCore('10.0')) dist = Math.abs(drawing.x+drawing.document.shape.width/2-data.x) + Math.abs(drawing.y+drawing.document.shape.height/2-data.y);
+            else dist = Math.abs(drawing.x+drawing.data.width/2-data.x) + Math.abs(drawing.y+drawing.data.height/2-data.y);
             if (dist < canvas.dimensions.size*5 && dist < nearestDistance) {
                 nearestDrawing = drawing;
                 nearestDistance = dist;